feat: split into packages
