+++ src/main/java/Remy/ChatbotException.java
@@ -1,3 +1,5 @@
+package Remy;
+
 public class ChatbotException extends Exception {
     public ChatbotException(String message) {
         super(message);
