+++ src/main/java/Remy/command/DeadlineCommand.java
@@ -1,3 +1,10 @@
+package Remy.command;
+
+import Remy.task.TaskList;
+import Remy.Ui;
+import Remy.Storage;
+import Remy.ChatbotException;
+import Remy.task.Deadline;
 import java.time.format.DateTimeParseException;
 
 public class DeadlineCommand extends Command {
