+++ src/main/java/Remy/command/DeleteCommand.java
@@ -1,14 +1,20 @@
-public class UnmarkCommand extends Command {
-    private int index;
-    public static final String COMMAND_WORD = "unmark";
+package Remy.command;
 
-    public UnmarkCommand(String input) throws ChatbotException {
+import Remy.task.TaskList;
+import Remy.Ui;
+import Remy.Storage;
+import Remy.ChatbotException;
+
+public class DeleteCommand extends Command {
+    private int index;
+    public static final String COMMAND_WORD = "delete";
+
+    public DeleteCommand(String input) throws ChatbotException {
         if (input.length() < 8) throw new ChatbotException("missing info lah.");
         int index = Integer.parseInt(input.substring(7)) - 1;
         if (index >= 0) {
             this.index = index;
         }
-
     }
 
     @Override
@@ -18,12 +24,9 @@ public class UnmarkCommand extends Command {
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws ChatbotException {
-
-        if (this.index >= taskList.size()) {
-            throw new ChatbotException("No such item lah.");
-        }
-        taskList.get(index).markAsUndone();
-        String content = "Done. You happy?\n" + taskList.get(index).toString();
+        String task = taskList.get(this.index).toString();
+        taskList.remove(this.index);
+        String content = "Done. Can you don't be so troublesome?\n" + task;
         storage.save(taskList);
         Ui.printShortSandwich(content);
     }
