+++ src/main/java/Remy/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package Remy.command;
+
+import Remy.task.TaskList;
+import Remy.Ui;
+import Remy.Storage;
+
 public class ExitCommand extends Command {
     public static final String COMMAND_WORD = "exit";
