+++ src/main/java/Remy/task/Event.java
@@ -1,3 +1,5 @@
+package Remy.task;
+
 public class Event extends Task {
 
     protected String startTime;
