+++ src/main/java/Remy/task/Task.java
@@ -1,3 +1,5 @@
+package Remy.task;
+
 import java.io.Serializable;
 
 public class Task implements Serializable {
