+++ text-ui-test/EXPECTED.TXT
@@ -1,6 +1,6 @@
 ____________________________________________________________
 
-I'm Remy, and it is NOT nice to see you.
+I'm Remy.Remy, and it is NOT nice to see you.
 Faster tell me what you want and go away.
 ____________________________________________________________
